[CoreBundle] Fixing test managers.

+++ Manager/WorkspaceManager.php
@@ -244,9 +244,12 @@ class WorkspaceManager
     public function deleteWorkspace(AbstractWorkspace $workspace)
     {
         $root = $this->resourceManager->getWorkspaceRoot($workspace);
+        $children = $root->getChildren();
 
-        foreach ($root->getChildren() as $node) {
-            $this->resourceManager->delete($node);
+        if ($children) {
+            foreach ($children as $node) {
+                $this->resourceManager->delete($node);
+            }
         }
         $this->om->remove($workspace);
         $this->om->flush();
